#2 modularizing a function
